dagster_code/pyproject.toml
[project]
name = "dagster_code"
requires-python = ">=3.10,<3.14"
version = "0.1.0"
dependencies = [
    "dagster==1.12.6",
    "dagster-dbt>=0.28.6",
    "dagster-duckdb>=0.28.6",
    "dbt-duckdb>=1.10.0",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
force-include = { "pyproject.toml" = "pyproject.toml" }

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "dagster_code"
registry_modules = [
    "dagster_code.components.*",
]


